+++ synthetic_well_logs/student_answers/__init__.py
@@ -1,5 +1,5 @@
 from synthetic_well_logs.student_answers.checker import check_student_answer
 from synthetic_well_logs.student_answers.reports import StudentCheckReport
-from synthetic_well_logs.student_answers.schema import StudentAnswer
+from synthetic_well_logs.student_answers.schema import FaciesInterval, StudentAnswer
 
-__all__ = ["StudentAnswer", "StudentCheckReport", "check_student_answer"]
+__all__ = ["FaciesInterval", "StudentAnswer", "StudentCheckReport", "check_student_answer"]